♻️ Refactor: LinkSubFolderUtils 컴포넌트 분리 및 이름 재지정
