+++ src/containers/folder-page/sub-folder-list/LinkSubFolderList.style.tsx
@@ -11,26 +11,6 @@ export const SubFolderBtnList = styled.div`
   }
 `;
 
-export const SubFolderUtilList = styled.div`
-  display: flex;
-  gap: 12px;
-`;
-
-export const Button = styled.button`
-  border: none;
-  color: var(--lb-gray60);
-
-  display: flex;
-  align-items: center;
-  text-align: center;
-  gap: 4px;
-  background-color: var(--lb-white);
-
-  @media (max-width: 767px) {
-    padding: 0;
-  }
-`;
-
 export const SubFolderBtn = styled.button<{ $state: boolean }>`
   color: var(--lb-black);
   font-family: Pretendard;
