+++ src/containers/folder-page/sub-folder-util/LinkSubFolderUtils.style.tsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const SubFolderUtilList = styled.div`
+  display: flex;
+  gap: 12px;
+`;
+
+export const Button = styled.button`
+  border: none;
+  color: var(--lb-gray60);
+
+  display: flex;
+  align-items: center;
+  text-align: center;
+  gap: 4px;
+  background-color: var(--lb-white);
+
+  @media (max-width: 767px) {
+    padding: 0;
+  }
+`;
